Use `module_function`
This converts the series of `def self.x` and `private_class_method`
calls to something a bit easier to read.

Unfortunately, Rubocop had something to say about that. It wanted us to
use the naked `private` declaration that's more common in Ruby code, but
it doesn't let `module_function` methods call them that way. But Rubocop
also doesn't like using `extend self`, so honestly this is the only way
to make that work.

I don't understand Rubocop's objection to inline `private`, so maybe
it's a better idea to disable that linter rule since we didn't opt into
it to begin with.