Add an x to ensure the random id string doesn't end in digits
The YoutubeTag::VALID_ID_REGEX will accept any number of
digits (optionally including an h mark for hours). This causes tests
passing a random string of characters like "aaabbbcccdd123123123" to
be recognized (with time_parameter set to 123123123 in the match
data).

Ensure an invalidating character (x) is added to the end of the string
to prevent mistaking the digits past the 11th as a time parameter.

It would appear there is a required but missing "?" character, or a
reasonable upper bound on hours, or a requirement that the h character
be passed when sending a time offset.