Maybe click the same focused selection?
I saw some errors during local testing (and I assume this find, focus, find
again, click pattern exists to work around other errors) where the
focus worked, but the find by role after focus, before click, was
giving errors.

My first attempt was to just chain the calls:

```javascript
cy.findByRole('option', { name: /@search_user_3/ }).focus().click();
```

However, this was failing as before (it gave a more interesting error,
that the parent of the focused element was not visible, so it could
not be clicked.

I didn't understand this problem yet, so attempted to work around it
by calling cy.focused().click()

```javascript
cy.findByRole('option', {name: /@search_user_3/ }).focus();
cy.focused().click();
```

This continued to fail when the viewport didn't include the parent
textbox, so the listbox became invisible/hidden when the focus was
activated.

I tried to scroll back up to the top of the comment area before
proceeding, by scrolling to the Discussion heading, but this also
didn't resolve the issue.

I tried to scroll the focused item into view, but this also didn't
work reliably, and shows the same errors.

```javascript
cy.findByRole('option', { name: /@search_user_3/ }).focus();
cy.focused().scrollIntoView().click();
```

I need to find a way to scroll up _after_ the search text activates
and displays the options, and before we focus, so that the textbox is
still in focus and doesn't mangle.

My assumption is that this _might_ happen with an adaptive
technology or possibly even just keyboard manipulation rather than
using the mouse, I can play with it a bit later.